re-organized modules and imports (#1068)
