+++ src/gluonts/nursery/anomaly_detection/supervised_metrics/_buffered_precision_recall.py
