+++ src/gluonts/nursery/anomaly_detection/supervised_metrics/_segment_precision_recall.py
